# lambda/openapi.json
{"openapi": "3.0.3", "info": {"title": "Powertools API", "version": "1.0.0"}, "servers": [{"url": "/"}], "paths": {"/create_todo": {"post": {"summary": "POST /create_todo", "description": "Creates a todo item in dynamodb by passing todo id, todo name and todo description and status", "operationId": "create_todo_create_todo_post", "parameters": [{"description": "title of the todo", "required": true, "schema": {"type": "string", "title": "Title", "description": "title of the todo"}, "name": "title", "in": "query"}, {"description": "todo description", "required": true, "schema": {"type": "string", "title": "Description", "description": "todo description"}, "name": "description", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "boolean", "title": "Return", "description": "Whether the todo was created successfully"}}}}}}}, "/list_todos": {"get": {"summary": "GET /list_todos", "description": "list all todos in the dynamodb table", "operationId": "list_todos_list_todos_get", "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {}, "type": "array", "title": "Return", "description": "A list of todos from the dyanmodb table"}}}}}}}, "/update_todo": {"put": {"summary": "PUT /update_todo", "description": "Update a todo in dynamodb table based on its todo id,title,description and status ", "operationId": "update_todo_update_todo_put", "parameters": [{"description": "Id of todo item to be updated", "required": true, "schema": {"type": "string", "title": "Todo Id", "description": "Id of todo item to be updated"}, "name": "todo_id", "in": "query"}, {"description": "title of todo", "required": true, "schema": {"type": "string", "title": "Title", "description": "title of todo"}, "name": "title", "in": "query"}, {"description": "Description of todo", "required": true, "schema": {"type": "string", "title": "Description", "description": "Description of todo"}, "name": "description", "in": "query"}, {"description": "Status of tod", "required": true, "schema": {"type": "string", "title": "Status", "description": "Status of tod"}, "name": "status", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "boolean", "title": "Return", "description": "Whether todo was updated successfully"}}}}}}}, "/delete_todo": {"delete": {"summary": "DELETE /delete_todo", "description": "delete a todo item in dynamodb table based on its id", "operationId": "delete_todo_delete_todo_delete", "parameters": [{"description": "Id of todo item to be updated", "required": true, "schema": {"type": "string", "title": "Todo Id", "description": "Id of todo item to be updated"}, "name": "todo_id", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "boolean", "title": "Return", "description": "Whether todo was deleted successfully"}}}}}}}, "/get_todo": {"get": {"summary": "GET /get_todo", "description": "get a todo in dynamodb table based on its id", "operationId": "get_todo_get_todo_get", "parameters": [{"description": "Id of todo item to be updated", "required": true, "schema": {"type": "string", "title": "Todo Id", "description": "Id of todo item to be updated"}, "name": "todo_id", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "string", "title": "Return", "description": "retrieve a todo"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}